+++ packages/core/src/components/mark-bar/mark-bar.interfaces.ts
@@ -1,0 +1,12 @@
+export interface MarkBarInputEvent {
+  originalEvent: Event;
+  value: string;
+}
+
+export interface MarkBarPaginationEvent {
+  originalEvent: MouseEvent;
+}
+
+export interface MarkBarClearEvent {
+  originalEvent: MouseEvent;
+}