+++ packages/dso-toolkit/src/components/navbar/navbar.args.ts
@@ -1,11 +1,14 @@
 import { ArgTypes } from "@storybook/types";
 
-import { Navbar, NavbarItem } from "./navbar.models.js";
+import { NavBarToggleExtensionEvent, Navbar, NavbarItem } from "./navbar.models.js";
 
 export interface NavbarArgs {
   items: NavbarItem[];
   modifier: string;
   open: boolean;
+  extensionOpen?: boolean;
+  extensionAnimation?: boolean;
+  dsoExtensionToggle(event: NavBarToggleExtensionEvent): void;
 }
 
 export const navbarArgTypes: ArgTypes<NavbarArgs> = {
@@ -25,12 +28,32 @@ export const navbarArgTypes: ArgTypes<NavbarArgs> = {
       type: "boolean",
     },
   },
+  extensionOpen: {
+    control: {
+      type: "boolean",
+    },
+  },
+  extensionAnimation: {
+    control: {
+      type: "boolean",
+    },
+  },
+  dsoExtensionToggle: {
+    action: "dsoExtensionToggle",
+  },
 };
 
-export function navbarArgsMapper(a: NavbarArgs): Navbar {
+export function navbarArgsMapper<TemplateFnReturnType>(
+  a: NavbarArgs,
+  extension?: TemplateFnReturnType
+): Navbar<TemplateFnReturnType> {
   return {
     items: a.items,
     modifier: a.modifier,
     open: a.open,
+    extension,
+    extensionOpen: !!extension && a.extensionOpen,
+    extensionAnimation: !!extension && !!a.extensionAnimation,
+    dsoExtensionToggle: (e) => a.dsoExtensionToggle(e),
   };
 }