+++ storybook/src/components/mark-bar/mark-bar.core-stories.ts
@@ -1,0 +1,26 @@
+import type { Meta } from "@storybook/web-components";
+
+import { templateContainer } from "../../templates";
+
+import readme from "@dso-toolkit/core/src/components/alert/readme.md?raw";
+import { MarkBarArgs, markBarMeta, markBarStories } from "dso-toolkit";
+
+const meta: Meta<MarkBarArgs> = {
+  ...markBarMeta({ readme }),
+  title: "Core/Mark Bar",
+};
+
+export default meta;
+
+const { Default } = markBarStories({
+  templateContainer,
+  storyTemplates: (templates) => {
+    const { markBarTemplate } = templates;
+
+    return {
+      markBarTemplate,
+    };
+  },
+});
+
+export { Default };